+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-## [60.1.0] 
+## [60.2.0] 
 - Resources was updated from DIPS.Mobile.DesignTokens
+
+## [60.1.1]
+- [ScrollView][Android] Fixed `HasAdditionalSpaceAtTheEnd` not applying bottom spacing when consumer sets `Padding` on the ScrollView.
+
+## [60.1.0]
+- [DateAndTimePicker] `MaximumDate` and `MinimumDate` now also constrain the time portion when the selected date equals the boundary date.
+- [TimePicker] Added `MinimumTime` and `MaximumTime` properties to constrain selectable times.
 
 ## [60.0.4]
 - [ContextMenu] Added `ContextMenuItem.IsEnabled` bindable property to disable individual context menu items. Disabled items appear grayed out and cannot be tapped.